MDT-70: fix - changePw page
