+++ medit-frontend/src/locales/translations.json
@@ -1,74 +1,82 @@
 {
   "en": {
-    "login": "Login",
-    "logout": "Logout",
-    "signup": "Sign Up",
-    "registration": "Registration",
-    "continueWithoutRegistration": "Continue without registration",
-    "firstName": "First name",
-    "lastName": "Last name",
-    "email": "Email",
-    "confrim": "Confirm",
-    "confirmPassword": "Confirm password",
-    "patient": "Patient",
-    "whoAreYouQuestion": "Who are you?",
-    "getStarted": "Get started",
-    "startSlogan": "Your digital healthcare assistant",
-    "scanQrCode": "Scan QR code",
-    "scanQrCodeDescription": "Scan the patient's QR for access",
-    "qrCode": "QR code",
-    "patientQrCodeDescritpion": "Caregivers can connect to the patient's profile by scanning this code from the app",
-    "scanMe": "Scan me",
-    "home": "Home",
-    "profile": "Profile",
-    "userInfo": "user information",
-    "support": "Support",
-    "settings": "Settings",
-    "patientList": "patient list",
+    "add": "add",
     "caregiver": "Caregiver",
     "changePassword": "change password",
-    "selectLanguage ": "select language",
-    "phone": "Phone",
+    "confirmNewPassword": "confirm new password",
+    "confirmPassword": "Confirm password",
+    "confrim": "Confirm",
     "contactForm": "Contact form",
+    "continueWithoutRegistration": "Continue without registration",
+    "currentPassword": "current password",
+    "email": "Email",
     "FAQs": "FAQs",
-    "writeAMessage": "Write a message ...",
+    "firstName": "First name",
+    "getStarted": "Get started",
+    "home": "Home",
+    "lastName": "Last name",
+    "login": "Login",
+    "logout": "Logout",
+    "newPassword": "new password",
+    "patient": "Patient",
+    "patientList": "patient list",
     "patientListDescription": "Select a patient profile or add one",
-    "add": "add"
+    "patientQrCodeDescritpion": "Caregivers can connect to the patient's profile by scanning this code from the app",
+    "phone": "Phone",
+    "profile": "Profile",
+    "qrCode": "QR code",
+    "registration": "Registration",
+    "save": "save",
+    "scanMe": "Scan me",
+    "scanQrCode": "Scan QR code",
+    "scanQrCodeDescription": "Scan the patient's QR for access",
+    "selectLanguage ": "select language",
+    "settings": "Settings",
+    "signup": "Sign Up",
+    "startSlogan": "Your digital healthcare assistant",
+    "support": "Support",
+    "userInfo": "user information",
+    "whoAreYouQuestion": "Who are you?",
+    "writeAMessage": "Write a message ..."
   },
   "it": {
+    "add": "aggiungi",
+    "caregiver": "Caregiver",
+    "cambiaPassword": "cambia password",
+    "confermaNuovaPassword": "conferma nuova password",
+    "confermaPassword": "Conferma password",
+    "confrim": "Conferma",
+    "contactForm": "Modulo di contatto",
+    "continueWithoutRegistration": "Continua senza registrazione",
+    "currentPassword": "password attuale",
+    "email": "Email",
+    "FAQs": "FAQs",
+    "firstName": "Nome",
+    "getStarted": "Inizia",
+    "home": "Home",
+    "lastName": "Cognome",
     "login": "Accedi",
     "logout": "Esci",
-    "signup": "Registrati",
-    "registration": "Registrazione",
-    "continueWithoutRegistration": "Continua senza registrazione",
-    "firstName": "Nome",
-    "lastName": "Cognome",
-    "email": "Email",
-    "confrim": "Conferma",
-    "confirmPassword": "Conferma password",
+    "newPassword": "nuova password",
     "patient": "Paziente",
-    "whoAreYouQuestion": "Chi sei?",
-    "getStarted": "Inizia",
-    "startSlogan": "Il tuo assistente digitale per la salute",
+    "patientList": "lista pazienti",
+    "patientListDescription": "Seleziona un profilo paziente o aggiungine uno",
+    "patientQrCodeDescritpion": "I caregiver possono connettersi al profilo del paziente scansionando questo codice dall'app",
+    "phone": "Telefono",
+    "profile": "Profilo",
+    "qrCode": "Codice QR",
+    "registration": "Registrazione",
+    "save": "salva",
+    "scanMe": "Scansionami",
     "scanQrCode": "Scansiona QR code",
     "scanQrCodeDescription": "Scansiona il QR del paziente per accedere",
-    "qrCode": "Codice QR",
-    "patientQrCodeDescritpion": "I caregiver possono connettersi al profilo del paziente scansionando questo codice dall'app",
-    "scanMe": "Scansionami",
-    "home": "Home",
-    "profile": "Profilo",
-    "userInfo": "informazioni utente",
-    "support": "Supporto",
-    "settings": "Impostazioni",
-    "patientList": "lista pazienti",
-    "caregiver": "Caregiver",
-    "changePassword": "cambia password",
     "selectLanguage": "seleziona lingua",
-    "phone": "Telefono",
-    "contactForm": "Modulo di contatto",
-    "FAQs": "FAQs",
-    "writeAMessage": "Scrivi un messaggio ...",
-    "patientListDescription": "Seleziona un profilo paziente o aggiungine uno",
-    "add": "aggiungi"
+    "settings": "Impostazioni",
+    "signup": "Registrati",
+    "startSlogan": "Il tuo assistente digitale per la salute",
+    "support": "Supporto",
+    "userInfo": "informazioni utente",
+    "whoAreYouQuestion": "Chi sei?",
+    "writeAMessage": "Scrivi un messaggio ..."
   }
 }
